view-app/view/requirements.txt
flask==2.3.0
mysql-connector-python-rf
